HTTP client moved to Core
