+++ platform-examples/simple-rest/simple-rest-client/src/main/java/com/canoo/dolphin/samples/rest/RestClient.java
@@ -16,8 +16,8 @@
 package com.canoo.dolphin.samples.rest;
 
 import com.canoo.platform.client.PlatformClient;
-import com.canoo.platform.client.http.HttpClient;
-import com.canoo.platform.client.http.RequestMethod;
+import com.canoo.platform.core.http.HttpClient;
+import com.canoo.platform.core.http.RequestMethod;
 
 public class RestClient {
