+++ platform/dolphin-platform-client/src/main/java/com/canoo/dp/impl/platform/client/http/DefaultHttpURLConnectionResponseHandler.java
@@ -15,7 +15,7 @@
  */
 package com.canoo.dp.impl.platform.client.http;
 
-import com.canoo.platform.client.http.HttpURLConnectionHandler;
+import com.canoo.platform.core.http.HttpURLConnectionHandler;
 
 import java.net.HttpURLConnection;
