+++ platform/dolphin-platform-client/src/main/java/com/canoo/dp/impl/platform/client/http/HttpClientProvider.java
@@ -3,10 +3,10 @@ package com.canoo.dp.impl.platform.client.http;
 import com.canoo.dp.impl.platform.client.AbstractServiceProvider;
 import com.canoo.platform.client.ClientConfiguration;
 import com.canoo.platform.client.PlatformClient;
-import com.canoo.platform.client.http.HttpClient;
-import com.canoo.platform.client.http.HttpURLConnectionFactory;
-import com.canoo.platform.client.http.spi.RequestHandlerProvider;
-import com.canoo.platform.client.http.spi.ResponseHandlerProvider;
+import com.canoo.platform.core.http.HttpClient;
+import com.canoo.platform.core.http.HttpURLConnectionFactory;
+import com.canoo.platform.core.http.spi.RequestHandlerProvider;
+import com.canoo.platform.core.http.spi.ResponseHandlerProvider;
 import com.google.gson.Gson;
 
 import java.util.Iterator;
