+++ platform/dolphin-platform-client/src/main/java/com/canoo/dp/impl/platform/client/http/HttpRequestImpl.java
@@ -3,10 +3,10 @@ package com.canoo.dp.impl.platform.client.http;
 import com.canoo.dp.impl.platform.core.Assert;
 import com.canoo.dp.impl.platform.core.PlatformConstants;
 import com.canoo.platform.client.ClientConfiguration;
-import com.canoo.platform.client.http.ByteArrayProvider;
-import com.canoo.platform.client.http.HttpRequest;
-import com.canoo.platform.client.http.HttpResponse;
-import com.canoo.platform.client.http.HttpURLConnectionHandler;
+import com.canoo.platform.core.http.ByteArrayProvider;
+import com.canoo.platform.core.http.HttpRequest;
+import com.canoo.platform.core.http.HttpResponse;
+import com.canoo.platform.core.http.HttpURLConnectionHandler;
 import com.canoo.platform.core.DolphinRuntimeException;
 import com.google.gson.Gson;
