+++ platform/dolphin-platform-client/src/main/java/com/canoo/dp/impl/platform/client/http/HttpResponseImpl.java
@@ -2,10 +2,10 @@ package com.canoo.dp.impl.platform.client.http;
 
 import com.canoo.dp.impl.platform.core.Assert;
 import com.canoo.platform.client.ClientConfiguration;
-import com.canoo.platform.client.http.ByteArrayProvider;
-import com.canoo.platform.client.http.HttpExecutor;
-import com.canoo.platform.client.http.HttpResponse;
-import com.canoo.platform.client.http.HttpURLConnectionHandler;
+import com.canoo.platform.core.http.ByteArrayProvider;
+import com.canoo.platform.core.http.HttpExecutor;
+import com.canoo.platform.core.http.HttpResponse;
+import com.canoo.platform.core.http.HttpURLConnectionHandler;
 import com.canoo.platform.core.DolphinRuntimeException;
 import com.google.gson.Gson;
