--- platform/dolphin-platform-client/src/main/java/com/canoo/platform/client/http/spi/ResponseHandlerProvider.java
@@ -1,9 +1,0 @@
-package com.canoo.platform.client.http.spi;
-
-import com.canoo.platform.client.ClientConfiguration;
-import com.canoo.platform.client.http.HttpURLConnectionHandler;
-
-public interface ResponseHandlerProvider {
-
-    HttpURLConnectionHandler getHandler(ClientConfiguration configuration);
-}
