+++ platform/dolphin-platform-client/src/main/resources/META-INF/services/com.canoo.platform.core.http.spi.ResponseHandlerProvider
