+++ platform/dolphin-platform-core/src/main/java/com/canoo/platform/core/domain/UrlToAppDomainConverter.java
@@ -1,4 +1,4 @@
-package com.canoo.platform.client.domain;
+package com.canoo.platform.core.domain;
 
 import java.net.URL;
