+++ platform/dolphin-platform-core/src/main/java/com/canoo/platform/core/http/ByteArrayProvider.java
@@ -1,4 +1,4 @@
-package com.canoo.platform.client.http;
+package com.canoo.platform.core.http;
 
 public interface ByteArrayProvider {
