+++ platform/dolphin-platform-core/src/main/java/com/canoo/platform/core/http/spi/RequestHandlerProvider.java
@@ -1,0 +1,9 @@
+package com.canoo.platform.core.http.spi;
+
+import com.canoo.platform.core.PlatformConfiguration;
+import com.canoo.platform.core.http.HttpURLConnectionHandler;
+
+public interface RequestHandlerProvider {
+
+    HttpURLConnectionHandler getHandler(PlatformConfiguration configuration);
+}
