+++ platform/dolphin-platform-remoting-client/src/main/java/com/canoo/dp/impl/client/ClientContextFactoryImpl.java
@@ -23,8 +23,8 @@ import com.canoo.dp.impl.remoting.codec.OptimizedJsonCodec;
 import com.canoo.dp.impl.remoting.legacy.util.Function;
 import com.canoo.platform.client.ClientConfiguration;
 import com.canoo.platform.client.PlatformClient;
-import com.canoo.platform.client.http.HttpClient;
-import com.canoo.platform.client.http.HttpURLConnectionHandler;
+import com.canoo.platform.core.http.HttpClient;
+import com.canoo.platform.core.http.HttpURLConnectionHandler;
 import com.canoo.platform.client.session.ClientSessionStore;
 import com.canoo.platform.remoting.DolphinRemotingException;
 import com.canoo.platform.remoting.client.ClientContext;
