+++ platform/dolphin-platform-remoting-client/src/main/java/com/canoo/dp/impl/client/DolphinPlatformHttpClientConnector.java
@@ -23,8 +23,8 @@ import com.canoo.dp.impl.remoting.commands.DestroyContextCommand;
 import com.canoo.dp.impl.remoting.legacy.communication.Codec;
 import com.canoo.dp.impl.remoting.legacy.communication.Command;
 import com.canoo.platform.client.ClientConfiguration;
-import com.canoo.platform.client.http.HttpClient;
-import com.canoo.platform.client.http.RequestMethod;
+import com.canoo.platform.core.http.HttpClient;
+import com.canoo.platform.core.http.RequestMethod;
 import com.canoo.platform.remoting.DolphinRemotingException;
 import com.canoo.platform.remoting.client.RemotingExceptionHandler;
 import org.slf4j.Logger;
