+++ platform/dolphin-platform-remoting-client/src/test/java/com/canoo/dolphin/client/impl/TestDolphinPlatformHttpClientConnector.java
@@ -30,8 +30,8 @@ import com.canoo.dp.impl.remoting.legacy.communication.JsonCodec;
 import com.canoo.dp.impl.remoting.legacy.util.Provider;
 import com.canoo.platform.client.HeadlessToolkit;
 import com.canoo.platform.client.PlatformClient;
-import com.canoo.platform.client.http.HttpClient;
-import com.canoo.platform.client.http.HttpURLConnectionFactory;
+import com.canoo.platform.core.http.HttpClient;
+import com.canoo.platform.core.http.HttpURLConnectionFactory;
 import com.canoo.platform.remoting.DolphinRemotingException;
 import com.google.gson.Gson;
 import org.testng.Assert;
